refactor(traits)!: move HasPaginationParams from Http\Concerns to Traits namespace
BREAKING CHANGE: HasPaginationParams trait namespace changed from
App\Http\Concerns to App\Traits. Also adds getSortParams() helper
and normalizes empty strings to null in extractDbFilters().

Migration: Update all use statements from:
- OLD: use App\Http\Concerns\HasPaginationParams;
- NEW: use App\Traits\HasPaginationParams;
